添加了音量条处理。

+++ 3rdparty/include/libavutil/common.h
@@ -168,7 +168,7 @@ static av_always_inline av_const int16_t av_clip_int16_c(int a)
 static av_always_inline av_const int32_t av_clipl_int32_c(int64_t a)
 {
     if ((a+0x80000000u) & ~UINT64_C(0xFFFFFFFF)) return (a>>63) ^ 0x7FFFFFFF;
-    else                                         return (int32_t)a;
+    else                                         return a;
 }
 
 /**
